+++ src/main/java/com/groupfour/chatapp/chatapp/controllers/FileAttachmentController.java
@@ -1,0 +1,16 @@
+package com.groupfour.chatapp.chatapp.controllers;
+
+import com.groupfour.chatapp.chatapp.services.FileAttachmentService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class FileAttachmentController {
+
+    private FileAttachmentService fileAttachmentService;
+
+    @Autowired
+    public FileAttachmentController(FileAttachmentService fileAttachmentService) {
+        this.fileAttachmentService = fileAttachmentService;
+    }
+}
