+++ src/main/java/com/groupfour/chatapp/chatapp/controllers/OptionController.java
@@ -1,0 +1,20 @@
+package com.groupfour.chatapp.chatapp.controllers;
+
+
+import com.groupfour.chatapp.chatapp.services.OptionService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping(value="/polls")
+public class OptionController {
+
+    private OptionService optionService;
+
+    @Autowired
+    public OptionController(OptionService optionService) {
+        this.optionService = optionService;
+    }
+
+
+}
