+++ src/main/java/com/groupfour/chatapp/chatapp/controllers/VoteController.java
@@ -1,0 +1,46 @@
+package com.groupfour.chatapp.chatapp.controllers;
+
+
+import com.groupfour.chatapp.chatapp.models.Vote;
+import com.groupfour.chatapp.chatapp.repositories.VoteRepository;
+import com.groupfour.chatapp.chatapp.services.VoteService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+
+@RestController
+public class VoteController {
+
+    private VoteRepository voteRepository;
+    private VoteService voteService;
+
+    @Autowired
+    public VoteController(VoteRepository voteRepository) {
+        this.voteRepository = voteRepository;
+    }
+
+    @RequestMapping(value = "/polls/{pollId}/votes", method = RequestMethod.POST)
+    public ResponseEntity<?> createVote(@PathVariable Long pollId, @RequestBody Vote vote) {
+        vote = voteRepository.save(vote);
+        // Set the headers for the newly created resource
+        HttpHeaders responseHeaders = new HttpHeaders();
+        responseHeaders.setLocation(ServletUriComponentsBuilder.
+                fromCurrentRequest().path("/{id}").buildAndExpand(vote.getVoteId()).toUri());
+        return new ResponseEntity<>(null, responseHeaders, HttpStatus.CREATED);
+    }
+
+    @RequestMapping(value="/polls/votes", method=RequestMethod.GET)
+    public Iterable<Vote> getAllVotes() {
+        return voteRepository.findAll();
+    }
+
+    @RequestMapping(value="/polls/{pollId}/votes", method=RequestMethod.GET)
+    public Iterable<Vote> getVote(@PathVariable Long pollId) {
+        return voteRepository.findVotesByPoll(pollId);
+    }
+
+
+}
