--- src/main/java/com/groupfour/chatapp/chatapp/fileAttachment/FileAttachmentController.java
@@ -1,15 +1,0 @@
-package com.groupfour.chatapp.chatapp.fileAttachment;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class FileAttachmentController {
-
-    private FileAttachmentService fileAttachmentService;
-
-    @Autowired
-    public FileAttachmentController(FileAttachmentService fileAttachmentService) {
-        this.fileAttachmentService = fileAttachmentService;
-    }
-}
