--- src/main/java/com/groupfour/chatapp/chatapp/fileAttachment/FileAttachmentRepository.java
@@ -1,8 +1,0 @@
-package com.groupfour.chatapp.chatapp.fileAttachment;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface FileAttachmentRepository extends CrudRepository<FileAttachment, Long> {
-}
