--- src/main/java/com/groupfour/chatapp/chatapp/fileAttachment/FileAttachmentService.java
@@ -1,16 +1,0 @@
-package com.groupfour.chatapp.chatapp.fileAttachment;
-
-import com.groupfour.chatapp.chatapp.fileAttachment.FileAttachmentRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class FileAttachmentService {
-
-    private FileAttachmentRepository fileAttachmentRepository;
-
-    @Autowired
-    public FileAttachmentService(FileAttachmentRepository fileAttachmentRepository) {
-        this.fileAttachmentRepository = fileAttachmentRepository;
-    }
-}
