+++ src/main/java/com/groupfour/chatapp/chatapp/models/User.java
@@ -1,0 +1,93 @@
+package com.groupfour.chatapp.chatapp.models;
+
+
+
+import com.groupfour.chatapp.chatapp.models.Message;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import java.util.Date;
+import java.util.Set;
+
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue
+    private Long userId;
+
+    private String userName;
+    private String password;
+    private String email;
+
+    private Date timeStamp = new Date();
+    private Integer activeStatus = 0;
+
+    @OneToMany
+    private Set<Message> messages;
+
+
+    public User() {
+    }
+
+    public User(String userName, String password, String email) {
+        this.userName = userName;
+        this.password = password;
+        this.email = email;
+    }
+
+
+    public Long getUserId() {
+        return userId;
+    }
+
+    public String getUserName() {
+        return userName;
+    }
+
+    public void setUserName(String userName) {
+        this.userName = userName;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public Date getTimeStamp() {
+        return timeStamp;
+    }
+
+    public void setTimeStamp(Date timeStamp) {
+        this.timeStamp = timeStamp;
+    }
+
+    public Set<Message> getMessages() {
+        return messages;
+    }
+
+    public void setMessages(Set<Message> messages) {
+        this.messages = messages;
+    }
+
+    public Integer getActiveStatus() {
+        return activeStatus;
+    }
+
+    public void setActiveStatus(Integer activeStatus) {
+        this.activeStatus = activeStatus;
+    }
+}
