--- src/main/java/com/groupfour/chatapp/chatapp/poll/OptionController.java
@@ -1,21 +1,0 @@
-package com.groupfour.chatapp.chatapp.poll;
-
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping(value="/polls")
-public class OptionController {
-
-    private OptionService optionService;
-
-    @Autowired
-    public OptionController(OptionService optionService) {
-        this.optionService = optionService;
-    }
-
-
-}
