--- src/main/java/com/groupfour/chatapp/chatapp/poll/OptionRepository.java
@@ -1,13 +1,0 @@
-package com.groupfour.chatapp.chatapp.poll;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface OptionRepository extends CrudRepository<Option, Long> {
-
-
-
-
-
-}
