--- src/main/java/com/groupfour/chatapp/chatapp/poll/PollRepository.java
@@ -1,19 +1,0 @@
-package com.groupfour.chatapp.chatapp.poll;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface PollRepository extends CrudRepository<Poll, Long> {
-
-    @Query(value = "SELECT v.* " +
-            "FROM Option o, Vote v " +
-            "WHERE o.POLL_ID = ?1 " +
-            "AND v.OPTION_ID = o.OPTION_ID", nativeQuery = true)
-Poll findPollByChatAndPoll(Long id,Long pollId);
-
-Iterable<Poll> findPollsByChat(Long id);
-
-
-}
