--- src/main/java/com/groupfour/chatapp/chatapp/poll/VoteRepository.java
@@ -1,15 +1,0 @@
-package com.groupfour.chatapp.chatapp.poll;
-
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface VoteRepository extends CrudRepository<Vote, Long> {
-    @Query(value = "SELECT v.* " +
-            "FROM Option o, Vote v " +
-            "WHERE o.POLL_ID = ?1 " +
-            "AND v.OPTION_ID = o.OPTION_ID", nativeQuery = true)
-    public Iterable<Vote> findVotesByPoll(Long pollId);
-}
-
