--- src/main/java/com/groupfour/chatapp/chatapp/poll/VoteService.java
@@ -1,15 +1,0 @@
-package com.groupfour.chatapp.chatapp.poll;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class VoteService {
-
-    private VoteRepository voteRepository;
-
-    @Autowired
-    public VoteService(VoteRepository voteRepository) {
-        this.voteRepository = voteRepository;
-    }
-}
