+++ src/main/java/com/groupfour/chatapp/chatapp/repositories/ChatRepository.java
@@ -1,0 +1,15 @@
+package com.groupfour.chatapp.chatapp.repositories;
+
+import com.groupfour.chatapp.chatapp.models.Chat;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatRepository extends CrudRepository<Chat, Long>{
+    public Chat findByChatName(String chatName);
+
+
+
+
+
+}
