+++ src/main/java/com/groupfour/chatapp/chatapp/repositories/MessageRepository.java
@@ -1,0 +1,11 @@
+package com.groupfour.chatapp.chatapp.repositories;
+
+import com.groupfour.chatapp.chatapp.models.Message;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface MessageRepository extends CrudRepository<Message, Long> {
+    Iterable<Message> findMessagesBySender(Long userId);
+    Iterable<Message> findMessagesByDestinationChat(Long chatId);
+}
