+++ src/main/java/com/groupfour/chatapp/chatapp/repositories/UserRepository.java
@@ -1,0 +1,12 @@
+package com.groupfour.chatapp.chatapp.repositories;
+
+import com.groupfour.chatapp.chatapp.models.User;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends CrudRepository<User, Long> {
+
+
+
+}
