+++ src/main/java/com/groupfour/chatapp/chatapp/repositories/VoteRepository.java
@@ -1,0 +1,16 @@
+package com.groupfour.chatapp.chatapp.repositories;
+
+import com.groupfour.chatapp.chatapp.models.Vote;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface VoteRepository extends CrudRepository<Vote, Long> {
+    @Query(value = "SELECT v.* " +
+            "FROM Option o, Vote v " +
+            "WHERE o.POLL_ID = ?1 " +
+            "AND v.OPTION_ID = o.OPTION_ID", nativeQuery = true)
+    public Iterable<Vote> findVotesByPoll(Long pollId);
+}
+
