--- src/main/java/com/groupfour/chatapp/chatapp/user/UserRepository.java
@@ -1,11 +1,0 @@
-package com.groupfour.chatapp.chatapp.user;
-
-import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-
-
-
-}
