--- src/main/java/com/groupfour/chatapp/chatapp/user/UserService.java
@@ -1,14 +1,0 @@
-package com.groupfour.chatapp.chatapp.user;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserService {
-    private UserRepository userRepository;
-
-    @Autowired
-    public UserService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-}
